refactored vizToolBar into seperate components for easier maintenance
