+++ src/components/vizToolbar/components/vizNav/vizNav.module.css
@@ -1,0 +1,15 @@
+.vizNav:not(:last-child) {
+  margin: 0;
+}
+
+@media screen and (max-width: 768px) {
+  .navBtn {
+    min-width: 100px;
+  }
+}
+
+@media screen and (min-width: 769px) {
+  .navBtn {
+    min-width: 150px;
+  }
+}
